ItEr15S04ArquitecturaServidorItEr14S04: improvement to "checkVersion" method.
Now the "checkVersion" method also checks if the version of the entity passed as a parameter to "checkVersion" is null. If so, the check is considered to be sucsessfull (because an entity with version == null is considered a new object, that is, an object which does not exist in the database).