ItEr18S04ArquitecturaServidorItEr17S04: IGenericDao::findExistingEntity + several minor improvements.
The method IGenericDao::findExistingEntity has been added. Unlike IGenericDao::find, it returns a runtime exception if the entity does not exist. So, this method should be used when the entity with the key passed as a parameter is supposed to exist. This patch also fixes some methods in GenericDaoHibernate which were not using try-catch to convert Hibernate exceptions to the Spring exception hierarchy.