ItEr44S14CUAsignarUsuarioAProxectoPlanificacionItEr43S19: Implemented restriction on edition of orders.
Now only authorized users or those with the role EDIT_ALL_ORDERS will
be able to edit an order. The remaining users will be able to open the
edition interface but they will find the save buttons disabled.
TODO: disable all the fields in the edition interface so they can't be
modified.