ItEr46S13CUImportacionRecursosProductivosItEr45S11: Added support for updating when importing criterion types.
Support for updating criterion types when importing criterion types has been added.

To update a criterion type, it is only necessary to send the modifications, and not all the state. In consequence, it is only necessary to send modified propert
ies in the criterions and the criterion type itself, and only new or modified cr
iterions need to be sent. Furthermore, any legal change in the criterion hierarchy of the criterion type is allowed.

Finally, a test case has been added to test the updating funcionality.