ItEr47S04ValidacionEProbasFuncionaisItEr46S04: [Bug #282] Fixing bug.
The problem lies in that the start and end data were retrieved from
allocation. Allocation start and end are calculated looking into day
assignments. The first day was assigned to zero hours when applying
the splines and this caused that the allocation start slipped to the
next day. As the first interval end was the task's next day too it
provoked the failure.