ItEr49S04ValidacionEProbasFuncionaisItEr48S04: Improving allocation correctness.
Taking into account the availability of the resources when
distributing some hours in a interval of dates. This way if at a day
no hours can be assigned because it doesn't match the criteria or
it's a holiday for the resource no hours are assigned. If in the
interval there are no days in which could be assigned hours, no
assignments are created.