ItEr57S04ValidacionEProbasFuncionaisItEr56S04: [Bug #432] Fix bug.
The start date is always the start of the task. It was happening that
the first days of the allocation there were not assignable resources,
so no day assignments were generated for these first days. Resource
allocation start date was incorrectly calculated looking at the day
assignments. Now it always return the start of the task.