ItEr57S15RecalculosConexionEscenarios: Fix bug.
The bug happens when moving a task if the final position is the same
as before moving it. In that case the task is left in the position
onto which is moved instead of the correct position(the initial
position). This happens because a no event is produced, old value is
equal to new value. Checking for this on TaskComponent.