ItEr57S04ValidacionEProbasFuncionaisItEr56S04: [Bug #463] Fix bug.
When filtering at resource load screen a new TimeTrackerComponent is
created and the old one is discarded. The old one remains in memory,
and listens to zoom change event from TimeTracker. Because of already
not being on the page it fails when being recreated as shown in the
bug's stack trace.