ItEr57S04ValidacionEProbasFuncionaisItEr56S04: [Bug #489] Remove spurious containers.
This fixes the bug for the planning view. It removes the containers
for the current scenario at all SchedulingDataForVersions but the new
current one so their day assignments don't interfere when calculating
the external load.

The newly created SchedulingDataForVersion has also spurious
containers resulting from the deep copy. These are removed too.

Now the data is stored correctly so when you get back after saving the
load is shown correctly. But while the planning is being edited the
chart shows wrong data.