ItEr57S04ValidacionEProbasFuncionaisItEr56S04: [Bug #489] Fix bug.
ReturningNewAssignments was always adding the initial new assignments
(assignments for the current order and scenario that are still not
saved). But removals and addings can happen on these new assignments
but ReturningNewAssignments kept using the initial new assignments.

Instead of passing the initial new assignmetns to
ReturningNewAssignments, they are added to resource using @AfterCopy
mechanism.