ItEr60S04ValidacionEProbasFuncionaisItEr59S04: [Bug #519] Fix bug.
Removing inverse. When a set is mapped with inverse=true it modifies
the association based on the other side of the association, in this
case the advanceAssignment property at AdvanceMeasurement. The
advanceAssignment property at AdvanceMeasurement was correctly set at
null and the set has cascade all-delete-orphan. This could make you
think that it should remove the AdvanceMeasurement. But in the moment
that Hibernate tries to apply the cascade the AdvanceMeasurement has
been removed from the set so it doesn't try to apply the changes.