ItEr60S04ValidacionEProbasFuncionaisItEr59S04: [Bug #519] Fix bug.
Changing the set instance was interfering with delete-orphan. The
initial persistent set must be kept in order to delete-orphan keep on
working. If a new set is assigned hibernate loses track of the
removals done and doesn't delete the orphans. Instead of assigning to
new set, the advanceMeasurements set is reset to the new contents.