ItEr60S04ValidacionEProbasFuncionaisItEr59S04: [Bug #570] Fix bug.
Now it removes the tasks that would be removed to avoid being
re-saved. When starting the edition the tasks must be loaded,
otherwise it would fail when accessing the collections for removing
the task of the order element being removed. Their dependencies must
be loaded since the removal of associated dependencies is necessary.