Move order.code constraint must be unique from entity to model
There was a constraint in Order that checked the code of an order was
unique in the whole system. This validation was done for each Order,
which slowed the process when saving. This validation is now moved to
OrderModel.save. It does the same funcionality but faster since now it's
no needed to retrieve all order elements for each order.

FEA: ItEr60S18CambiosPantallaArbolPedido