Move constraint validation hoursgroup.code must be unique to OrderModel.save()
HoursGroup.code must be unique for each OrderElement. This constraint
was being validation in HoursGroup, as an Hibernate validation. This
validation was very slow has for each HoursGroup it was needed to check
against all other HoursGroup. I moved this validation to
OrderModel.save() and speed up the process, keeping the same
funcionality.

FEA: ItEr61S07PerformanceOrderEdition