Fix broken build, zkex dependency not found.
The zkex dependency was moved to professional edition ZK repository in
version 3.6.4. This move must have been done after our migration to
3.6.4 from 3.6.3. This error happens when doing a maven
install on a machine that hasn't dowloaded the zkex dependency
previously.

The solution is to add professional edition ZK repository. Taking a
look at the source code of zkex it turns out it's GPL 3 licensed (at
least for 3.6.4) so this is safe.

FEA: ItEr60S04ValidacionEProbasFuncionaisItEr59S04