Adds extra cache of class validators in HibernateValidator.ClassValidator
The method getClassValidator() in HibernateValidator.ClassValidator tries
to find in a Map a previously created ClassValidator. In case it doesn't
exist, the method instances a new ClassValidator but doesn't cache it.

The operation of creating a new ClassValidator is expensive, 3 ms
aproximately. When an Order is saved this method is executed many times
(around 20000 times). Validators for objects of type ArrayList, PersistenSet,
Collections$UnmodifiableCollection are created each time. As an outcome,
the process of validating is remarkably slow (around 10 seconds).

This patch adds a new class NavalPlanClassValidator, which is a verbatim
copy of HibernateValidatorClassValidator, but it adds an extra cache for
ClassValidators created in getClassValidator().

FEA: ItEr60S18CambiosPantallaArbolPedido