Fix bug
It was returning the resources satisfaying the criterions at the very
same moment of the query instead of returning the resources that
satisfy the criteria at some point in time. The former is wrong
because if in the future the resource satisfies the criteria, it can
be used in the allocation.

FEA: ItEr60S04ValidacionEProbasFuncionaisItEr59S04