Fix problem when adding extra hours at the end date of a partially completed date
If you add extra hours to a partially completed last day and you quit
the same extra hours to another day the resources per day count
changes and this is confusing for the user. The implemented solution
is to keep the same IntraDayDate end when modifying the last day.

The real solution for the underlying would be to keep an spec object
with the desired values from the user so when having to reallocate
uses the original intention of the user and not the current resources
per day and total hours values.

FEA: ItEr61S05BugFixing