Handle scheduling strategy when initalizing startDate of Tasks
In many parts of the code is assumed that tasks are always schedule
forward. It's assumed, for instance, that startDate of a Tasks is always
know and that its endDate can be infer from its number of hours
(considering initially a 8h/day effort). This logic is widely spread
in many parts of the code (specially classes like Task, TaskGroup,
OrderElement, etc).

Now with backwards planning it's necessary to detect this cases, as it
will be possible to the opposite operation, that means, knowning the
endDate of a Task infer its startDate.

This patch also modifies the behaviour of AS_LATE_AS_POSSIBLE constraint
for taking care of backwards-planning. As a result, now tasks are
correctly positioned when set AS_LATE_AS_POSSIBLE in backwards
scheduling strategy.

FEA: ItEr62OTS04PlanificacionHaciaAtras