[Bug #682] Fix bug
After saving the order in the order detail,
manageOrderElementAdvancesController is initialized and associated to
the order that is being saved. Then you go to the scheduling view
through a change of perspective and save. This causes the order to
increase its version number. If you return to the order detail view,
manageOrderElementAdvancesController keeps pointing to the former
Order. This is the reason that when executing its bussiness logic
fails with an org.hibernate.StaleObjectStateException.

Now when going back to the scheduling view the order is reloaded from
the DB and the controllers of each tab are set to null so they are
initialized on demand with the current Order instead of sticking to
the former one.

I think the problem was introduced at commit:
4bb7e1ffd43221c5af9c42357b9e19bcc2dd702c. The performance improvements
reported in that commit can be compromised by this.

FEA: ItEr62S05BugFixing