Replaced listener on zul code with listener on widget
* When GanttPannel's zoomLevel is changed we have to update the timetracker and scroll the pannel, so
we need the timetracker's scrollLeft value. Instead of asking for it to the GanttPannel after the listener
is executed at Planner.java we will send both the selected zoomLevel and the scroll value. We replaced the
listener at the zul file and added a new listener at Planner's widget class. Currently we are only sending the
desired zoomLevel

FEA: ItEr02S03MigracionZK5