The start and end date are always set when reseting all allocations
By default, if they are not specified, the start and end are taken
from the task. This allows to calculate resources per day correctly
when doing backwards scheduling. The external behavior is the same for
resource allocations' start date. For the end date now instead of
returning the date of the last day assignment, the date set at the
reset method would be returned.

FEA: ItEr62OTS04PlanificacionHaciaAtras