[Bug #702] Fix bug
The problem lied in that the end date of the task group was being
updated keeping the length when the start changed. Since the children
of the group were not modified when adding the milestone the end date
was not recalculated and the wrong value persisted.

Calculating the end date keeping the length in task groups is
incorrect since their end date doesn't depend on the previous length,
but in the incoming dependencies. Several refactorings and
simplifications applied thanks to this insight.

FEA: ItEr62S05BugFixing