Added new allocating function: Sigmoid Function
Adds a new type of allocating function.

This function allocates hours into resources following a Sigmoid
function.

Known bugs:

* The calendar of a resource (Specific) or the calendar of a Task
(Generic) is not being respected. For instance, the algorithm allocates
hours on bank holidays.

* By default, the algorithm only considers periods of time of 1 hour. If
the calculation estimates an allocation of 0.25 hours for one day, the
result if truncated to 0 hours. This usually happens at the beginning
and at the end of task. Solving the problem at the beginning is easy,
just need to skip allocations with 0 hours assigned, shifting to another
day. But this is not a totally satisfactory solution as it leaves gaps
by the end of the task. To solve this, one solution could be to increase
the period of time by one day at the end, till there won't be 0 hours
allocated on the last day.

FEA: ItEr62OTS04FuncionCurvaS