Added new menu tab for calculating MonteCarlo graph for an order
* Added configuration option for enabling this menu tab

* Montecarlo method calculates a range of enddates and its corresponding
likelihood. For each task of an order, there are three possible
durations (pessimistic, normal and optimistic). Each duration is
associated with a likelihood value. A montecarlo calculation, generates
for each task a random value and takes the duration of the task
associated to this random value, summing each value for each task, the
resuling value is the hypothetical enddate of that order. This operation
is executed n times. The higher the value, the more accurate the
results, although there is a threshold limit.

FEA: ItEr63OTS06MonteCarlo