Fix bug calculate multiple 'Critical paths' for MonteCarlo
It's necesary to take into account several issues when calculating all
the critical paths for a schedule:

The starting point should be a task with no incoming dependencies or
with incoming dependencies to other tasks which are not a 'Critical
path' task.

A task can navigate through its taskgroup, if any, to other tasks, but
it only navigates to those tasks which have no incoming dependencies or
have only incoming dependencies with other tasks which are not children
of that taskgroup.

FEA: ItEr63OTS06MonteCarlo