[Bug #729] Fix bug
The problem lies in that the startContraintDate was being written to
the start date of the task directly. So, the start date was set to the
specified date without the end date being recalculated. This could
leave the task in an inconsistent state, with an end date prior to the
start date.

The solution is to simply not update directly the task start date with
the constraint date.

FEA: ItEr63S03BugFixing