Now more tasks are reachable from a TaskPoint
Now from an end date change the start can change because of backwards
scheduling. This must be considered when considering if a dependency
provokes an infinite loop or when calculating the recalculations to be
done.

For example, if a end date of a tasks is modified the start date could
be moved and this could provoke other tasks depending on this start
date to be moved. This is true for all tasks but the containers that
have a special behavior.

FEA: ItEr64OTS03PlanificacionHaciaAtras