[Bug #750] Fix bug
The problem was that a milestone can be a top level task and doesn't
have an associated order element. If the parent of the top level task
is retrieved, the taskgroup associated to the order is returned and
from this taskgroup the order can always be obtained.

FEA: ItEr65S06BugFixing