[Bug #753] Fix bug
The parents recalculations were recalculating the secondary point,
this caused that a posterior recalculation for the secondary point of
the same task would say it has not modified the task. This caused the
depending recalculations to not be executed.

Now when doing the parent recalculation only the primary point is
modified.

FEA: ItEr65S06BugFixing