Change behaviour for appropriative and non-appropriative allocations
Now, when a queue element is allocated at a specific date:

* Unschedule all elements in target queue from that date on. If there's
already an element at that date, unschedule it too.
* Schedule those elements back in the queue in topological order. This
guarantees that dependencies are satisfied.

As a result, the new element is allocated at specified date and all the
elements after that day are shifted in the queue. If there were
dependencies between those shifted elements and other elements in the
queue, they will be shifted in their own queue too.

Take into account scheduling time when moving a task to a specific date.

Distinguish between how to generate day assigments between appropriate and
non-appropriative allocations.

FEA: ItEr66OTS03AlgoritmosLimitantesItEr65OTS05