[Bug #766] Fix bug
The problem lied in that the task's start date changed due to the task
being moved but the resource allocation keep on using its own date
instead of the new value. This leads to using a stale value for doing
the allocation. This can cause that the used start date was posterior
that the new end date as it happened in the exception.

FEA: ItEr66S04BugFixing