[Bug #760] Fix corner case
When the calculated value was end date and the task was moved
backwards with a new end date that is before to the old start date and
the allocation cannot be satisfied, the start date picked would be
posterior to the new end. To avoid this, the dates must always be
recalculated taking into account the length.

FEA: ItEr66S04BugFixing