Fix bug when allocating an unscheduled queue element, do it like if it were like an unscheduled element in the list on unscheduled queue elements
If doing an appropriative allocation it's necessary to unscheduled
element to make room enough for a gap, later those elements should be
schedule back again like it were unscheduled elements coming from the
list of unscheduled queue elements.

It's not possile to make any guess about in which queue it make fit,
as this may change depending on the new gaps created.

FEA: ItEr66OTS08CorreccionsRecursosLimitantesItEr65OTS04