Improve the stickiness of the parameters specified by the user
After doing an allocation if the task was moved to a place where it
cannot be satisfied the parameters specified by the user were
overriden with zero like values.

Now it keeps the fields originalTotalAssignment and resourcesPerDay so
if the allocation can be satisfied again it uses the original values.

FEA: ItEr66S04BugFixing