[Bug #786] Fix bug
The error was happening due to using Dates instead of LocalDates. The
start of the satisfaction was in the middle of a day and the received
Date was at the start, so the first day was considered outside of the
criterion satisfaction's interval. Criterion satisfactions now use
LocalDates instead of Dates and this problem is fixed.

FEA: ItEr67S04BugFixing