[Bug #861] Fix bug
Deleting the order elements without parents after the task sources
have been synchronized.

When an OrderElement is removed it's left behind whithout parent. The
method deleteOrderElementWithoutParent is responsible of deleting
them.  If this is done before doing the tasks synchronization, the
associated TaskGroup to the OrderGroup removed has as children the
moved task. This causes said task to be processed twice, from two
differnet possitions: at the OrderGroup removed and at the position it
was added.

Now deleteOrderElementWithoutParent it's done after the
synchronization, so the associated TaskGroup doesn't have the moved
task.

FEA: ItEr71S04BugFixing