Keep using the same AssignmentsAllocator
In order to take into consideration the previous resources used the
same AssignmentsAllocator must be used in an allocation. This implies
changes in the until filling allocation that was instantiating an
AssignmentsAllocator for each day. Now one is instantiated and used
for each day until allocating the hours.

FEA: ItEr71S07FragmentationDeletionItEr70S09