[Bug #932] Check there are not overlapped category assignments, and if there are mark failing criterio assignment in form
I refactored the code that does the checking in the business layer.
Instead of putting it directly in a method executed by
HibernateValidator I put it in a static method in the business layer.
This way it can be reused easily from the view layer.

The validation returns a ValidationException pointing to the object that
caused the error. Its corresponding component is localized in the view
and the error is shown as a WrongValidationException.

FEA: ItEr73S04BugFixing