[Bug #967] Transient instance editing task in WBS after set criterion and save
If a treeitem is selected and a new element is added, the selected item
is converted into a container (newObject) and the selected item is
promoted to a child od this new container.

The criteria of the original element is copied to new element
(container) and removed from the original element. This causes that the
child has a new IndirectCriterionRequirement, which is a new object.

The error happened when doing a reattachment of the Order being this new
IndirectCriterionRequirement a transient entity.

Finally, we have decided to change the business logic of how
criterionRequirements are copied when a treeitem is promoted to a
container. What happens now is that the original item retains its
criterionRequirement and the newObject(container) has empty
criterionRequirements.

FEA: ItEr74S04BugFixing