[Bug #975] Respect consolidated day assignments of a Resource Allocation when applying a Stretch function
When an advance function is applied to a resource allocation, in this
case a Stretch function, the function should respect the period of
consolidated day assignments if any.

This means that an advance function should only calculate hours for day
assigments considering the period of time in the resource allocation that
is free (has no day assignments).

The class ConsolidatedStretch represents an interval that encompasses the
consolidated day assignments of a resource allocation, if any. This interval
is infered from the resource allocation and is never persisted.

This interval is shown in the window of intervals as disabled (the user
cannot edit it). It can happen that the StretchFunction may have several
intervals defined, then the user consolidates some days, resulting
that some of the intervals previously defined are now outdated (by the
consolidated interval). In this case, an overlap error will happen if
the user tries to apply the function, and the user should remove by hand
those outdated intervals.

FEA: ItEr74S04BugFixing