[Bug #1028] Avoid bug
updateAndRefresh() causes ConcurrentModificationException, however it's
need to do an update() when the project is saved, the graph changes, etc

The concurrent modification happens if a resource is allocated in a task
and the save project. If the project is saved before and then the
resource is allocated and save, it's OK.

Use refresh() instead to avoid the bug.

FEA: ItEr74S04BugFixing