[Bug #1070] Fix problem in GanttDiagramGraph
getRecalculationsNeeded took a huge amount of time when tasks can be
reached from several paths in complex graphs. In the case a task is
reached from another path the recalculation for that task was removed
from the result and added to the pending queue. Thus it was guaranteed
that the result was a topological order, since recalculations would be
pushed to the end. But the recalculations dependent of the already
added were still in the pending queue and executed. This caused more
removals from the result and subsequent additions to the pending
queue.

Now a topological order is applied to the recalculations
calculated. For each task point a depth value is calculated. A
topological order is necessary, so a recalculation is executed after
all its predecessors.

FEA: ItEr74S04BugFixing