Revert "Fix bug"
The bug was present when introducing a dependency between containers,
the children were not affected by the movement of the parent. The
solution fixed this problem, but introduced others. For example when
adding a dependency from a top level task that is not a container to a
container.

This reverts commit 16862f28f4f0f590cab072bd145e2af2cbf1fffe.

FEA: ItEr74S04BugFixing