[Bug #1082] Fix bug
Making that the parent recalculations go after the normal ones. Thus
if a normal recalculation is needed for a parent, i.e. a container,
it's executed before the parent recalculation. Otherwise the parent
recalculation would apply the modifications and the normal
recalculation will think that no modifications have been done and its
successors would not be executed.

FEA: ItEr74S04BugFixing