Revert "Fix bug"
The bug was present when introducing a dependency between containers,
the children were not affected by the movement of the parent. The
solution fixed this problem, but introduced others. For example
when adding a dependency from a top level task that is not a
container to a container.

This reverts commit 48ba72511d1dd4b5da832c37679911a896202020.

FEA: ItEr74S04BugFixing