Ensure that the mouseover event for TaskComponent is executed
For resizable tasks the z-index was less than the value for the
dependency. This prevented the mouseover event of happening for
TaskComponent. Without this event happening, the end of the dependency
was not detected and adding a dependency was impossible. This could be
tested with resizable tasks, e.g. not allocated yet tasks.

FEA: ItEr75S08MigrationZK5