[Bug #1117] Fix issue changing I18nHelper in business.
Current I18nHelper in business was doing real translation of strings.
Then it was passing the string translated to server language to webapp
module and this module tries to translate the string again to end-user
language.

In order to avoid this I18nHelper in business is modified and now it
just provides a marker method for translatable strings. Real translation
is going to be done in webapp module.

Several files need to be modified due to this change.

FEA: ItEr75S04BugFixing