[Bug #1124] Fix problem when calculating assigned duration
When an interval of an allocation is considered and a day assignment
is within that interval, the day assignment must not be summed
completely. Only the part that is inside of the said interval.

In other words, intervals can start and end in the middle of a day,
so not all the effort of the DayAssignment must be considered.

FEA: ItEr75S04BugFixing