[Bug #1132] Create new enum CalendarExceptionTypeColor
The enum is being used in CalendarExceptionType and all the entities
related have been updated to work with the new enum. Web service has
also been updated.

Colors defined now are: DEFAULT (red), GREEN and BLUE. Colors to show in
the interface are just examples and should be changed in a later patch.

FEA: ItEr75S04BugFixing