[Bug #820] Several changes in watermark height calculations
Fixed bug and aded some modifications to the system to calculate watermark heigth. An
enhancement to this could be calling ganttz.Planner.getInstance().adjustScrollableDimensions()
also on window resizes to avoid the appearance of a white area when increasing the backport
and before a recalculation is called.

FEA: ItEr75S04BugFixing