[Bug #1232] Fix problem in load chart filtering all assignments taking into account order resources
The problem was that it was using all the assignments to calculate
external load and overload. However, for capability line it was using
just the resources assigned each day.

Then I've added a new method to filter all the assignments taking into
account the order resources assigned each day.

FEA: ItEr75S04BugFixing