[Bug #1238] Fix concurrent modification exception
The fix consists of changing the persistence model of the Dependency entity.
So far, it was deleted in the database when orphan dependencies appear in the Hibernate session
being flushed. Now, the orphan dependencies are removed explicitely on saving the Planning state.
An orphan dependency in the database is the one which has origin or destination NULL.

FEA: ItEr75S04BugFixing