[Bug #1141] Remove Task.reloadResourcesTextIfChange().
We think it's not necessary to reload the resources text when the start date of
a task is updated, and that's what was being done here.

This change caused a invalidate() on the ZK object of the task, which destroyed
the old object and replaced it with a new one. Because of that, the setter we
implemented to fix bug #1141 didn't work.

FEA: ItEr75S04BugFixing