Removed problematic left and right listeners on TaskDetails treeRows
There was an issue while editing the cells inside TaskDetails component, as using the cursor keys
for moving inside the textboxes was causing that the cell was unfocused. The standard Tab and Shift+Tab
navigation is kept for right and left, and the arrow and down keys are kept mapped.

FEA: ItEr75S04BugFixing