[Bug #719] Fix problem in WBS when you modify tasks hours with parents in previous pages
The problem was related with a JavaScript error trying to update the
hours in the parent, when that element was not visible (in a previous
page).

A workaround is implemented to allow call directly
Intbox::setValueDirectly in order to prevent call smartUpdate which
launches the JavaScript event and the problem.

It is also implemented a method to know if Intbox is or not in current
page.

FEA: ItEr75S04BugFixing