[Bug #1262] Remove option to cancel progress assignments in Gantt view
It makes this window coherent with the one in WBS.

Moreover, there were more problems than the one described in the bug
itself. For example, if you remove a progress type and click cancel, the
progress is removed when you come back to the pop-up.

The problem is that we're not working against a copy so in cancel button
we need to revert all the changes done, which is quite complicated. So,
at this moment the decision is remove this option.

FEA: ItEr75S04BugFixing