doctool: Support translation of underlines for sub-pages
For files which do not qualify to be included as part of the HTML index
it is needed to translate the sectioning when assembling the TeX code
for PDF file generation, because in this case the files are included
inline and it is needed for them to create subsections instead of top
level chapters.

Also, the sorting is improved to take only into account the numnbering
that precedes the descriptive text in the filenames of .rst sources.
This makes sorting the files for inlining in the TeX output work as
expected.