Add vertical borders in watermarks to improve differentiation between days
When using day level, the vertical positioning inside the day was not very clear.
This also happened during consecutive inactivity periods, where all marked bank
holidays were displayed together. A border is added to the columns to avoid this
behaviour.

It would be useful to achieve an optimum level of alignment to review the pixel spacing
that appears on Chrome, and improve the rounding calculations and spacings between 1st
and 2nd level detail.

FEA: ItEr75S04BugFixing