[Bug #1326] Fix issue calculating properly hours to allocate
The issue was introduced in commit a998367bec9d6cf02729fa6283466a1f933a6e2b
while trying to fix a different problem in interpolation function.

The patch has been reverted and both issues fixed now.

Once the patch was reverted, the number of remaining hours was wrong calculated
as if it was 0 it returns the allocated hours instead of 0.

FEA: ItEr76S04BugFixing