Bug #1327: Fix issue changing method to do reassignments in consolidation process
The problem was that ResourceAllocation.DayAssignmentsState has a wrong start
day for consolidated tasks.

During the consolidation the method used to do the reassignment called
ResourceAllocation.allocateTheWholeAllocation which changed DayAssignmentsState
start date to first day not consolidated. Now this method is avoided, and the
start date in DayAssignmentsState is right.

FEA: ItEr76S04BugFixing