Bug #1342: Show save button in Gantt view when it is disabled
Add new method to ICommand to allow specify when a command is disabled or not.
In a disabled command the method doAction is not going to be called when you
click in the button.

Set ISaveCommand to disabled when user doesn't have permissions to save the
order or the order is marked as STORED.

FEA: ItEr76S04BugFixing