Bug #1354: Fix bug
Although the constraint is not applied it must be provided. Otherwise
the dominating forces would not take it into account. Normally this
wasn't a problem since if the secondary point date changes it was
provided. But it can happen that the secondary point doesn't change,
in that case the constraint must still be taken into account.