Created ITemplatesModel as a super interface for IOrderTemplatesModel and IBudgetTemplatesModel.
Now we will be able to use any implementation of the model (for OrderTemplates
or BudgetTemplates) in related components without distintion, while having two
different Spring beans for each model.

If you create two instances of the same interface, the @Autowired annotation is
unable to know which one of the instances it should use, that's why
BudgetTemplatesModel can't implement IOrderTemplatesModel directly.