Bug #1409: Fix problem replacing allocations for the ones related to the criterion
Without this patch, the replacement was been doing wrong as it has been
replacing the allocation related to a criterion with allocations related to any
of the criteria being shown.

This didn't happen in company view because of there's not unsaved changes, so
the replacement method do nothing.

FEA: ItEr76S04BugFixing