Configure a custom authentication filter
For the moment the same behavior than before the authentication filter is kept.
Later it will be used to do some redirects depending on if the user is bound or
not to any resource.

In order to define a custom authentication filter it's needed to set
auto-config="false" see
http://static.springsource.org/spring-security/site/docs/2.0.x/reference/ns-config.html#ns-auto-config

We also need an entry point specified by "entry-point-ref" attribute.

FEA: ItEr76S28UserDashboard