Fix bug: NullPointerException when opening Dashboard view
Dashboard view needs the CriticalPath. The path should be calculated
before opening the view, no inside the view as this could result into a
NullPointerException sometimes (the user should have opened the Gantt
view before). The very same error happened in MonteCarlo and it was
fixed.

The solution is to use CriticalPathBuilder to construct the
CriticalPath.

FEA: ItEr76S15OrganizingPerProjectDashboard